Returning...
Haven't been working on this for a while. I'm working on my while loop problem (I have an idea of the problem) as well as another problem I found. As of right now, I'm working on implementing a card class and will rewrite the code to work with the new class.